videos/ovation: Add OVATION model
The OVATION (Oval Variation, Assessment, Tracking, Intensity, and
Online Nowcasting) model is an empirical model of the intensity of the
aurora developed at the Johns Hopkins University, Applied Physics
Laboratory by Patrick Newell and co-workers. It provides a short-term
(30 to 90 minute) forecast of the location and intensity of the aurora.

Signed-off-by: Stefan Kühnel <git@stefankuehnel.com>